# Conflicts:
#	src/components/header/DropdownUserMenuNav.tsx

+++ app/api/posters/delete/route.ts
@@ -1,4 +1,5 @@
 import { NextResponse } from "next/server";
+import { UserRole } from "@prisma/client";
 import { getServerAuthSession } from "~/server/auth";
 import { prisma } from "~/server/db/prisma";
 
@@ -12,12 +13,36 @@ export async function POST(request: Request) {
   try {
     const { posterId } = await request.json();
 
-    const data = await prisma.poster.delete({
+    const poster = await prisma.poster.findUnique({
       where: {
         id: posterId,
       },
     });
-    return NextResponse.json({ status: 200, data });
+
+    if (!poster) {
+      return NextResponse.json({
+        status: 404,
+        message: "Poster not found",
+      });
+    }
+
+    if (
+      session.user.role !== UserRole.ADMIN &&
+      session.user.id !== poster.userId
+    ) {
+      return NextResponse.json({
+        status: 401,
+        message: "Not enough permission to do this",
+      });
+    }
+
+    await prisma.poster.delete({
+      where: {
+        id: posterId,
+      },
+    });
+
+    return NextResponse.json({ status: 204 });
   } catch {
     return NextResponse.json({ status: 500, message: "Internal Server Error" });
   }
